+++ cli/src/cli.rs
@@ -27,10 +27,10 @@ use solana_sdk::{
     fee_calculator::FeeCalculator,
     hash::Hash,
     instruction::InstructionError,
-    instruction_processor_utils::DecodeError,
     loader_instruction,
     message::Message,
     native_token::lamports_to_sol,
+    program_utils::DecodeError,
     pubkey::Pubkey,
     signature::{keypair_from_seed, Keypair, KeypairUtil, Signature},
     system_instruction::{create_address_with_seed, SystemError, MAX_ADDRESS_SEED_LEN},
