+++ programs/btc_spv/src/spv_processor.rs
@@ -8,7 +8,7 @@ use hex;
 use log::*;
 use solana_sdk::account::KeyedAccount;
 use solana_sdk::instruction::InstructionError;
-use solana_sdk::instruction_processor_utils::limited_deserialize;
+use solana_sdk::program_utils::limited_deserialize;
 use solana_sdk::pubkey::Pubkey;
 
 pub struct SpvProcessor {}
