+++ programs/config/src/config_processor.rs
@@ -5,7 +5,7 @@ use bincode::deserialize;
 use log::*;
 use solana_sdk::account::KeyedAccount;
 use solana_sdk::instruction::InstructionError;
-use solana_sdk::instruction_processor_utils::{limited_deserialize, next_keyed_account};
+use solana_sdk::program_utils::{limited_deserialize, next_keyed_account};
 use solana_sdk::pubkey::Pubkey;
 
 pub fn process_instruction(
