+++ runtime/src/nonce_utils.rs
@@ -3,8 +3,8 @@ use solana_sdk::{
     account_utils::StateMut,
     hash::Hash,
     instruction::CompiledInstruction,
-    instruction_processor_utils::limited_deserialize,
     nonce_state::NonceState,
+    program_utils::limited_deserialize,
     pubkey::Pubkey,
     system_instruction::SystemInstruction,
     system_program,
