+++ sdk/src/lib.rs
@@ -12,7 +12,6 @@ pub mod fee_calculator;
 pub mod hash;
 pub mod inflation;
 pub mod instruction;
-pub mod instruction_processor_utils;
 pub mod loader_instruction;
 pub mod message;
 pub mod move_loader;
@@ -21,6 +20,7 @@ pub mod native_token;
 pub mod nonce_state;
 pub mod packet;
 pub mod poh_config;
+pub mod program_utils;
 pub mod pubkey;
 pub mod rent;
 pub mod rpc_port;
